eval_framework/src/company_eval_framework/mcp/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "company-eval-framework"
version = "0.1.0"
description = "Evaluation framework for AI applications"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pyyaml>=6.0",
    "pydantic>=2.0",
]

[project.optional-dependencies]
mcp = [
    "mcp>=1.0.0",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0",
    "mypy>=1.0",
]

[project.scripts]
company-eval-mcp = "company_eval_framework.mcp.server:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
company_eval_framework = ["py.typed"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
